The locale file interface got revamped.
Instead of defining `\thesis@[name]` within the `[locale].def`
files and `\thesis@[name]En` within the `base.def`
files, each [locale] now defines macros in form
`\thesis@[locale]@[name]` within their `[locale].def` files meaning
that English is no longer special-cased and multiple locales can be
loaded at the same time. This enables the model, where style files
conditionally load additional locale files (like `czech` and
`english` in case of the style file of the Faculty of Science,
which was the original motivation for the change).

The string manipulation macros `\thesis@`, `\thesis@@`,
`\thesis@@upper`, `\thesis@@lower` and `\thesis@requireLocale`
macros were also added to alleviate the complexity added by this
model. More specifically, the style files need to be able to
require new locales (this functionality is provided by the
`\thesis@requireLocale` macros) and strings of the current locale
can no longer be accessed directly through a command sequence, but
a command sequence must be composed from letter tokens and from the
expansion of the `\thesis@locale` macro (this functionality is
provided by the `\thesis@@*` macros).

The `\thesis@season` macro now returns a string rather than a macro
to allow for easy command sequence composition using the `\thesis@`
macro.

The `\thesis@requireWithOptions` was renamed to
`\thesis@requireStyle` as it resembles the semantics of the
`\thesis@requireLocale` macro.

The Faculty of Science style file now correctly typesets the front
matter in Czech and English regardless of the main locale.

The implicit options of the `mu/base` style file are now `nolof`
and `nolot` instead of `lof` and `lot`, respectively.