Major locale and style interface changes.
`\thesis@selectLocale` is no longer performed globally for the
entire document during `\thesis@load` (effectively overriding the
user's hyphenation settings, if Babel's or Polyglossia's
`\languagename` differs from `\thesis@locale`). Instead,
`\thesis@preamble` and `\thesis@postamble` now locally switch the
locale and expand `\thesis@blocks@preamble` and
`\thesis@blocks@postamble`, which are the new redefinables. Since
this breaks the behaviour of `\thesis@blocks@mainMatter`, whose
effects would also be local, `\thesis@blocks@mainMatter` is now
executed directly by `\thesis@preamble` after closing the group and
becomes a new part of the interface between the class and the style
files.

As a preparation for the future inclusion of bibliography support,
the /csquotes/ package is loaded, `\thesis@<locale>@csquotesStyle`
is a new part of the locale interface that sets the /csquotes/
style of a locale, and `\thesis@selectLocale` now switches the
/csquotes/ style.